CustomException and update method

+++ src/main/java/com/gb/shop/service/ProductService.java
@@ -3,10 +3,12 @@ package com.gb.shop.service;
 import com.gb.shop.dao.ProductRepository;
 import com.gb.shop.dao.entity.Product;
 import com.gb.shop.dto.ProductDto;
+import com.gb.shop.exceptions.ProductException;
 import com.gb.shop.mapper.ProductMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 @Service
@@ -19,10 +21,14 @@ public class ProductService {
     }
 
     public ProductDto findById(UUID id) {
+
         return mapper.toProductDto(repository.findById(id).orElseThrow());
     }
 
     public void deleteById(UUID id) {
+        if (Objects.isNull(id)) {
+            throw new ProductException("Id not presented");
+        }
         repository.deleteById(id);
     }
 
@@ -36,5 +42,12 @@ public class ProductService {
 
     public List<ProductDto> findAll() {
         return mapper.toProductDtoList(repository.findAll());
+    }
+
+    public void updateProduct(UUID id, String newName, Double newPrice) {
+        Product product = repository.findById(id).orElseThrow();
+        product.setName(newName);
+        product.setPrice(newPrice);
+        repository.save(product);
     }
 }
